Extract Visual Rule Editor Component
